pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenticops"
version = "0.1.0"
description = "Agent-First Cloud Observability Platform for AWS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # AWS
    "boto3>=1.34.0",
    # Database
    "sqlalchemy>=2.0.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Web
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    # Agent/LLM
    "strands-agents>=0.1.0",
    # Utils
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.8.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "boto3-stubs[essential]>=1.34.0",
]

[project.scripts]
aiops = "agenticops.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agenticops"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
